requirements.txt
streamlit>=1.37.0
python-barcode>=0.15.1
pillow>=10.0.0
pandas>=2.0.0
reportlab

